check for existance

+++ database/migrations/2016_03_01_112837_change_desciption_to_text_type_in_meals_table.php
@@ -13,7 +13,9 @@ class ChangeDesciptionToTextTypeInMealsTable extends Migration
     public function up()
     {
         Schema::table('meals', function (Blueprint $table) {
-            $table->dropColumn('description');
+            if (Schema::hasColumn('meals', 'description')) {
+                $table->dropColumn('description');
+            }
         });
     }
